Create PRIVACY-POLICY.md
# Privacy Policy for Swiftie Trivia

Last Updated: December 17, 2023

## Overview

This privacy policy outlines the types of personal information that is received and collected by Swiftie Trivia and how it is used. We are committed to protecting the privacy of our users while providing a personalized and valuable service.

## Information Collection and Use

Swiftie Trivia does not collect any personal data from its users, with the exception of a voluntarily provided user name for leaderboard purposes. 

### Leaderboard

- When a user achieves a high score, they are given the option to be added to the Leaderboard.
- A screen is presented to the user with an option to Opt Out. If the user chooses not to Opt Out, they can enter a free-form user name.
- The user name, along with the user's score and the date, are saved to our server.
- This information is displayed on the app's Leaderboard screen.
- No additional personal data (such as email addresses or account information) is collected or stored.

## Data Storage and Sharing

- The data collected (user name, score, and date) is stored on a secure server.
- This information is only used for displaying on the Leaderboard within the Swiftie Trivia app.
- We do not share, sell, rent, or trade personally identifiable information with third parties for their promotional purposes.

## User Consent

By using Swiftie Trivia, users consent to the collection and use of their information as described in this privacy policy.

## Changes to This Privacy Policy

- Swiftie Trivia reserves the right to make changes to this privacy policy at any time by giving notice to its users on this page and possibly within Swiftie Trivia itself.
- Users are encouraged to frequently check this page for any changes.
- The date of the last modification is listed at the top of this page.

## Contact Us

If you have any questions regarding this privacy policy or the practices of this app, please contact us at SwiftieTrivia.gmail.com.